Remove pointer type conversion
Converting scalar pointers to integers breaks ops which translate pointers into tensors. These ops are removed during the lowering, but the verification error makes the lowering noisy. The ptr -> integer type conversion only happens when we convert the function arguments to GetArgVal ops, so we can handle the conversion outside the type converter. All other uses of pointers are deleted as part of respective op lowerings.